[JAVA-31831] Split-or-move-core-java-8-datetime-2 (#16325)
