+++ core-java-modules/core-java-8-datetime-3/src/test/java/com/baeldung/localdatetoiso/LocalDateToISOUnitTest.java
@@ -1,13 +1,14 @@
 package com.baeldung.localdatetoiso;
 
-import static org.junit.Assert.assertEquals;
-
 import org.junit.Test;
+
 import java.time.LocalDate;
+
+import static org.junit.Assert.assertEquals;
 
 public class LocalDateToISOUnitTest {
     @Test
-    void givenLocalDate_whenUsingDateTimeFormatter_thenISOFormat(){
+    public void givenLocalDate_whenUsingDateTimeFormatter_thenISOFormat(){
         LocalDateToISO localDateToISO = new LocalDateToISO();
         LocalDate localDate = LocalDate.of(2023, 11, 6);
 
@@ -17,7 +18,7 @@ public class LocalDateToISOUnitTest {
     }
 
     @Test
-    void givenLocalDate_whenUsingSimpleDateFormat_thenISOFormat(){
+    public void givenLocalDate_whenUsingSimpleDateFormat_thenISOFormat(){
         LocalDateToISO localDateToISO = new LocalDateToISO();
         LocalDate localDate = LocalDate.of(2023, 11, 6);
 
@@ -27,17 +28,18 @@ public class LocalDateToISOUnitTest {
     }
 
     @Test
-    void givenLocalDate_whenUsingJodaTime_thenISOFormat() {
+    public void givenLocalDate_whenUsingJodaTime_thenISOFormat() {
         LocalDateToISO localDateToISO = new LocalDateToISO();
         org.joda.time.LocalDate localDate = new org.joda.time.LocalDate(2023, 11, 6);
 
         String expected = "2023-11-06T00:00:00.000Z";
         String actual = localDateToISO.formatUsingJodaTime(localDate);
         assertEquals(expected, actual);
+        assertEquals(expected, actual);
     }
 
     @Test
-    void givenLocalDate_whenUsingApacheCommonsLang_thenISOFormat() {
+    public void givenLocalDate_whenUsingApacheCommonsLang_thenISOFormat() {
         LocalDateToISO localDateToISO = new LocalDateToISO();
         LocalDate localDate = LocalDate.of(2023, 11, 6);
